pyproject.toml
[project]
name = "rag-website"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "langchain-community==0.2.6",
    "groq>=0.9.0",
    "langchain-groq>=0.1.6",
    "bs4>=0.0.2",
    "tiktoken>=0.7.0",
    "chromadb>=0.5.5",
    "langchain-openai>=0.1.17",
    "langgraph>=0.1.14",
    "azure-identity>=1.17.1",
    "pydantic>=2.8.2",
    "langchain>=0.2.11",
    "fastembed==0.3.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/rag_website"]
